+++ src/Component/Class/NewClassPage.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import Container from "@material-ui/core/Container";
 import Button from '@material-ui/core/Button';
 import CssBaseline from "@material-ui/core/CssBaseline";
-import Material_RTL from "../Component/Material_RTL";
-import M_RTL from "../Component/M_RTL";
+import Material_RTL from "../Material_RTL";
+import M_RTL from "../M_RTL";
 import Grid from "@material-ui/core/Grid";
 import { ValidatorForm, TextValidator } from 'react-material-ui-form-validator';
 import InputAdornment from "@material-ui/core/InputAdornment";
