+++ src/Component/Class/insideClass.js
@@ -8,13 +8,7 @@ import Typography from '@material-ui/core/Typography';
 import IconButton from '@material-ui/core/IconButton';
 import Container from '@material-ui/core/Container';
 import Grid from '@material-ui/core/Grid';
-import Paper from '@material-ui/core/Paper';
-import Link from '@material-ui/core/Link';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import VerticalTabs from './../Component/TabMenuEditProfile' ;
-import NewClassDialog from '../Pages/NewClassPage' ;
-import ClassesPage from '../Pages/ClassesPage' ;
-import JoinClassDialog from '../Pages/JoinClass' ;
 // ------------------------
 import ListItem from '@material-ui/core/ListItem';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
