+++ src/Component/Question/QuestionBank.js
@@ -3,8 +3,8 @@ import { makeStyles, useTheme  } from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
 import Container from "@material-ui/core/Container";
 import CssBaseline from "@material-ui/core/CssBaseline";
-import Material_RTL from "../Component/Material_RTL";
-import M_RTL from "../Component/M_RTL";
+import Material_RTL from "../Material_RTL";
+import M_RTL from "../M_RTL";
 import Divider from '@material-ui/core/Divider';
 import InputLabel from '@material-ui/core/InputLabel';
 import MenuItem from '@material-ui/core/MenuItem';
