+++ src/Component/User/LogoutDialog.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import Button from '@material-ui/core/Button';
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
-import tokenConfig from '../utils/tokenConfig';
-import serverURL from '../utils/serverURL' ;
+import tokenConfig from '../../utils/tokenConfig';
+import serverURL from '../../utils/serverURL' ;
 import DialogTitle from '@material-ui/core/DialogTitle';
 import axios from 'axios' ;
