--- src/Request methods/getTokenfromLocalStorage.js
@@ -1,9 +1,0 @@
-import React from 'react' ;
-
-const token = () => {
-    const [value , setValue] = React.useState(
-        localStorage.getItem('token') || ''
-    );
-        
-    
-}
